Refactor duplicate headers from lib classes.
There were some Ruby warnings on a few lines where variables were
assigned and unused, so I did tweak some other stuff while I was in
there. Like a ternary whose condition returned true or false based on
the truth or falseness of the condition. That would be good if we were
getting paid by the byte.